% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_alu.R
\name{categorize_dfam}
\alias{categorize_dfam}
\title{Categorize ALU Sequences by Chromosome}
\usage{
categorize_dfam(hic_file_path, output_dir)
}
\arguments{
\item{hic_file_path}{Path to the Hi-C data file.}

\item{output_dir}{Directory where the categorized CSV files will be saved.}
}
\value{
Invisible NULL. The function is used for its side effects of reading, filtering, and writing data.
}
\description{
Reads ALU sequence data and categorizes them by chromosomes based on a given Hi-C data file.
}
