% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_alu.R
\name{clean_alu}
\alias{clean_alu}
\title{Main Function to Process ALU Data}
\usage{
clean_alu(hic_file_path, file_path, output_dir)
}
\arguments{
\item{hic_file_path}{Path to the Hi-C data file, used in categorizing ALU sequences by chromosome.}

\item{file_path}{Path to the DFAM data file, which contains ALU sequences to be processed.}

\item{output_dir}{Base directory where the processed output data will be saved.}
}
\value{
path of the folder where the final output is stored.
}
\description{
This is a pipeline function that sequentially processes ALU data from a DFAM file. 
It first reads and filters ALU sequences from a given DFAM file (read_dfam), 
then categorizes these sequences by chromosomes based on a Hi-C data file (categorize_dfam), 
and finally selects the ALU sequences with the highest probability for each position in the data frame 
(select_the_highest_probability_alu).
}
\details{
The `read_dfam` function reads a DFAM file in chunks, filters for sequences starting with 'ALU', and saves them to CSV files.
The `categorize_dfam` function reads the filtered ALU data and categorizes them by chromosomes based on the Hi-C data file.
The `select_the_highest_probability_alu` function selects the ALU sequences with the highest probability for each position.
}
