% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFromFitHIC.R
\name{graphFromFitHIC}
\alias{graphFromFitHIC}
\title{This function reads a TSV file containing Hi-C data and creates a graph list from it.}
\usage{
graphFromFitHIC(tsv_file, chromosome)
}
\arguments{
\item{tsv_file}{A character string specifying the path to the TSV file containing Hi-C data.}

\item{chromosome}{A character string specifying the chromosome to filter the data for.}
}
\value{
A list containing the edge information for the graph.
}
\description{
This function reads a TSV file containing Hi-C data and creates a graph list from it.
}
