% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hic_graph.R
\name{merge_hic_with_alu}
\alias{merge_hic_with_alu}
\title{Clean ALU Data from DFAM}
\usage{
merge_hic_with_alu(file_path, output_directory)
}
\arguments{
\item{output_directory}{String, path to the directory where the output files will be saved.}

\item{path_file}{String, path to the clean_alu temp data.}

\item{hic_data_file}{String, optional, path to the Hi-C data file.}

\item{alu_directory}{String, optional, path to the directory containing ALU data files.}

\item{atac_data_file}{String, optional, path to the ATAC-seq data file.}
}
\value{
NULL, The function operates through side effects (reading, processing, 
        and writing data) and does not return any value.
}
\description{
This function performs several steps to clean ALU data from DFAM. It reads the 'hg38.hits' file in chunks,
filters ALU sequences, writes them to CSV files, categorizes these sequences by chromosome,
and finally selects the ALU elements with the highest probability (minimum e-value) for each position.
}
\details{
The function assumes a specific directory structure and creates necessary directories if they do not exist.
}
