% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLeiden.R
\name{plotLeiden}
\alias{plotLeiden}
\title{Plot the output of the Leiden algorithm, coloring nodes based on community}
\usage{
plotLeiden(result, Alus = FALSE)
}
\arguments{
\item{result}{The result of the runLeiden function.}
}
\value{
A plot of the graph with nodes colored based on community.
}
\description{
Plot the output of the Leiden algorithm, coloring nodes based on community
}
