% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runLeiden}
\alias{runLeiden}
\title{Takes an R List, converts it to a Graph object, and runs the Leiden algorithm}
\usage{
runLeiden(graphList, iterations, gamma, theta, seed = 0L)
}
\arguments{
\item{graphList}{R List containing the graph}

\item{iterations}{number of iterations to run the Leiden algorithm}

\item{gamma}{resolution parameter}

\item{theta}{temperature parameter}
}
\value{
Rcpp::List : R List containing the graph, community assignments, and quality
}
\description{
Takes an R List, converts it to a Graph object, and runs the Leiden algorithm
}
